Struct numcodecs_python::PyCodec

pub struct PyCodec { /* private fields */ }

Wrapper around Python Codecs to use the Rust Codec API.

Implementations

impl PyCodec

pub fn from_registry_with_config<'de, D: Deserializer<'de>>( config: D, ) -> Result<Self, D::Error>

Instantiate a codec from the Registry with a serialized configuration.

The config must include the id field with the PyCodecClass::codec_id.

Errors

Errors if no codec with a matching id has been registered, or if constructing the codec fails.

pub fn from_codec(codec: Bound<'_, Codec>) -> Result<Self, PyErr>

Wraps a Codec to use the Rust Codec API.

Errors

Errors if the codecโs class does not provide an identifier.

pub fn as_codec<'py>(&self, py: Python<'py>) -> &Bound<'py, Codec>

Access the wrapped Codec to use its Python CodecMethods API.

pub fn into_codec(self, py: Python<'_>) -> Bound<'_, Codec>

Unwrap the Codec to use its Python CodecMethods API.