numcodecs_python

Struct PyCodecClassAdapter

pub struct PyCodecClassAdapter { /* private fields */ }

Wrapper around PyCodecClasses to use the DynCodecType API.

Implementations

impl PyCodecClassAdapter

pub fn from_codec_class(class: Bound<'_, PyCodecClass>) -> Result<Self, PyErr>

Wraps a PyCodecClass to use the DynCodecType API.

Errors

Errors if the codec class does not provide an identifier.

pub fn as_codec_class<'py>(&self, py: Python<'py>) -> &Bound<'py, PyCodecClass>

Access the wrapped PyCodecClass to use its PyCodecClassMethods API.

pub fn into_codec_class(self, py: Python<'_>) -> Bound<'_, PyCodecClass>

Unwrap the PyCodecClass to use its PyCodecClassMethods API.

pub fn with_downcast<T: DynCodecType, O: Ungil>( py: Python<'_>, class: &Bound<'_, PyCodecClass>, with: impl Send + Ungil + for<'a> FnOnce(&'a T) -> O, ) -> Option<O>

If class represents an exported DynCodecType T, i.e. it was initially created with crate::export_codec_class, the with closure provides access to the instance of type T.

If class is not an instance of T, the with closure is not run and None is returned.