numcodecs_python

Function export_codec_class

pub fn export_codec_class<'py, T: DynCodecType>(
    py: Python<'py>,
    ty: T,
    module: Borrowed<'_, 'py, PyModule>,
) -> Result<Bound<'py, PyCodecClass>, PyErr>

Export the DynCodecType ty to Python by generating a fresh PyCodecClass inside module and registering it with the PyCodecRegistry.

Errors

Errors if generating or exporting the fresh PyCodecClass fails.