Struct numcodecs_python::PyCodecAdapter

pub struct PyCodecAdapter { /* private fields */ }

Wrapper around PyCodecs to use the [Codec] API.

Implementations

impl PyCodecAdapter

pub fn from_registry_with_config<'de, D: Deserializer<'de>>( config: D, ) -> Result<Self, D::Error>

Instantiate a codec from the PyCodecRegistry with a serialized configuration.

The config must include the id field with the PyCodecClassMethods::codec_id.

Errors

Errors if no codec with a matching id has been registered, or if constructing the codec fails.

pub fn from_codec(codec: Bound<'_, PyCodec>) -> Result<Self, PyErr>

Wraps a PyCodec to use the [Codec] API.

Errors

Errors if the codecโs class does not provide an identifier.

pub fn as_codec<'py>(&self, py: Python<'py>) -> &Bound<'py, PyCodec>

Access the wrapped PyCodec to use its PyCodecMethods API.

pub fn into_codec(self, py: Python<'_>) -> Bound<'_, PyCodec>

Unwrap the PyCodec to use its PyCodecMethods API.

pub fn try_clone(&self, py: Python<'_>) -> Result<Self, PyErr>

Try to clone this codec.

Errors

Errors if extracting this codecโs config or creating a new codec from the config fails.