numcodecs_python

Trait PyCodecClassMethods

pub trait PyCodecClassMethods<'py>: Sealed {
    // Required methods
    fn codec_id(&self) -> Result<String, PyErr>;
    fn codec_from_config(
        &self,
        config: Borrowed<'_, 'py, PyDict>,
    ) -> Result<Bound<'py, PyCodec>, PyErr>;
    fn as_type(&self) -> &Bound<'py, PyType>;
}

Methods implemented for PyCodecClasses.

Required Methods

fn codec_id(&self) -> Result<String, PyErr>

Gets the codec identifier.

Errors

Errors if the codec does not provide an identifier.

fn codec_from_config( &self, config: Borrowed<'_, 'py, PyDict>, ) -> Result<Bound<'py, PyCodec>, PyErr>

Instantiate a codec from a configuration dictionary.

The config dict must not contain an id field.

Errors

Errors if constructing the codec fails.

fn as_type(&self) -> &Bound<'py, PyType>

Gets the PyType that this PyCodecClass represents.